MUZIMA-279 | Refactored to move activities to their right package.
