+++ src/main/java/com/muzima/view/cohort/CohortWizardActivity.java
@@ -37,6 +37,7 @@ import com.muzima.service.SntpService;
 import com.muzima.view.BroadcastListenerActivity;
 import com.muzima.view.CheckedLinearLayout;
 import com.muzima.view.HelpActivity;
+import com.muzima.view.forms.FormTemplateWizardActivity;
 import com.muzima.view.forms.MuzimaProgressDialog;
 
 import java.io.IOException;
