+++ src/main/java/com/muzima/view/concept/CustomConceptWizardActivity.java
@@ -6,7 +6,7 @@
  * that uses this code in a for-profit venture, please contact the copyright holder.
  */
 
-package com.muzima.view.cohort;
+package com.muzima.view.concept;
 
 import android.content.Context;
 import android.content.Intent;
@@ -26,6 +26,7 @@ import com.muzima.domain.Credentials;
 import com.muzima.service.MuzimaSyncService;
 import com.muzima.service.WizardFinishPreferenceService;
 import com.muzima.view.InstallBarCodeWizardActivity;
+import com.muzima.view.provider.CustomProviderWizardActivity;
 import com.muzima.view.forms.MuzimaProgressDialog;
 import com.muzima.view.preferences.ConceptPreferenceActivity;
