+++ src/main/java/com/muzima/view/location/LocationListActivity.java
@@ -1,19 +1,29 @@
-package com.muzima.view.cohort;
+/*
+ * Copyright (c) 2014. The Trustees of Indiana University.
+ *
+ * This version of the code is licensed under the MPL 2.0 Open Source license with additional
+ * healthcare disclaimer. If the user is an entity intending to commercialize any application
+ * that uses this code in a for-profit venture, please contact the copyright holder.
+ */
+
+package com.muzima.view.location;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.PowerManager;
 import android.view.View;
 import android.widget.Button;
+
 import com.actionbarsherlock.view.Menu;
 import com.muzima.R;
 import com.muzima.domain.Credentials;
+import com.muzima.view.provider.ProviderListActivity;
 import com.muzima.view.forms.MuzimaProgressDialog;
-import com.muzima.view.preferences.ProviderPreferenceActivity;
+import com.muzima.view.preferences.LocationPreferenceActivity;
 
 
-public class ProviderListActivity extends ProviderPreferenceActivity{
-    private static final String TAG = "ProviderListActivity";
+public class LocationListActivity extends LocationPreferenceActivity {
+    private static final String TAG = "LocationListActivity";
     private MuzimaProgressDialog muzimaProgressDialog;
     protected Credentials credentials;
     private boolean isProcessDialogOn = false;
@@ -29,9 +39,9 @@ public class ProviderListActivity extends ProviderPreferenceActivity{
         nextButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                navigateToNextActivity();
-            }
-        });
+                        navigateToNextActivity();
+                    }
+            });
     }
 
     @Override
@@ -44,17 +54,17 @@ public class ProviderListActivity extends ProviderPreferenceActivity{
     protected void onResume() {
         super.onResume();
         if(isProcessDialogOn){
-            turnOnProgressDialog("Downloading Provider(s)...");
+            turnOnProgressDialog("Downloading Observations and Encounters...");
         }
     }
 
     @Override
     protected int getContentView() {
-        return R.layout.activity_provider_list;
+        return R.layout.activity_location_list;
     }
 
     private void navigateToNextActivity() {
-        Intent intent = new Intent(getApplicationContext(), ConceptListActivity.class);
+        Intent intent = new Intent(getApplicationContext(), ProviderListActivity.class);
         startActivity(intent);
         finish();
     }
@@ -63,4 +73,5 @@ public class ProviderListActivity extends ProviderPreferenceActivity{
         muzimaProgressDialog.show(message);
         isProcessDialogOn = true;
     }
+
 }
